Change for loop to add divs depending on grid size
Remove the process of adding divs one after another. Now what will
happen is that we will create a grid of divs depending on the gridSize
binding. We create a new div(rowDiv) and start adding divs(newDiv) to
that div until we reach the gridSize value. Once that value is reached,
we create a new rowDiv and repeat until the amount of rowDivs reaches
the gridSize value.