Remove mainDiv after button press
Now when the button is pushed, the mainDiv will be removed from the
body and new mainDiv will be created and appened to the body with the
".main" class added to it as well. Doing this will make it so that if
there is already a grid on screen, it will remove that grid and create
a new grid with the new dimensions that were input from the prompt.